requirements.txt
# Core
transformers>=4.42
accelerate>=0.33
datasets>=2.20
peft>=0.11
bitsandbytes>=0.43

# NLP utils
spacy>=3.7
scikit-learn>=1.4
regex
nltk

# API / server
fastapi
uvicorn

# Data & DB
pydantic>=2.7
psycopg2-binary
SQLAlchemy>=2.0
pandas>=2.2
numpy>=1.26
pyyaml

# Testing
pytest
